agent_template_26786/requirements.txt
```
os
python-dotenv
langchain-nvidia-ai-endpoints
langgraph
langchain-community-tools
langchain-core
```